+++ flask-data/smarter/tests/test_breeds.py
@@ -244,6 +244,19 @@ class TestGetBreedList(AuthMixin, BaseCase):
         self.assertEqual(test, self.data[0])
         self.assertEqual(response.status_code, 200)
 
+    def test_get_breeds_unknown_arguments(self):
+        response = self.client.get(
+            self.test_endpoint,
+            headers=self.headers,
+            query_string={
+                'foo': 'bar',
+            }
+        )
+
+        self.assertEqual(response.status_code, 400)
+        self.assertEqual(
+            "Unknown arguments: foo", response.json['message'])
+
 
 class TestGetBreed(AuthMixin, BaseCase):
     fixtures = [